feat: correct work of show total cases component
